+++ .github/workflows/test.yml
@@ -1,14 +1,31 @@
 name: test
-on: push
+on: 
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
 jobs:
   test:
     runs-on: ubuntu-latest
     steps:
       - name: checkout repo
         uses: actions/checkout@v4
+        
       - name: Use Node.js
         uses: actions/setup-node@v4
         with:
           node-version: "18.x"
+          cache: "npm"
+      
       - run: npm ci
       - run: npm test
+
+      - name: Upload HTML coverage report
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+            name: coverage-report
+            path: coverage/lcov-report/
